==> 00-Scripts/07-Z-Scores for Mapping.ipynb <==
<h1>Create Z-Scores for Mapping</h1>

# For reproducibility
import random
import numpy as np
r_state = 42
random.seed(r_state) 
np.random.seed(r_state)

import os
import re
import pandas as pd
import seaborn as sns

import sklearn

from sklearn.preprocessing import scale
from sklearn import linear_model
from sklearn import tree
from sklearn import preprocessing
from sklearn import feature_selection
from sklearn import model_selection
from sklearn import metrics  
from sklearn import ensemble

#from sklearn.externals.six import StringIO
#from sklearn.model_selection import GridSearchCV
#from sklearn.feature_selection import SelectKBest 
#from sklearn.feature_selection import f_regression

from timeit import default_timer as timer
import datetime

# Needed on a Mac
import matplotlib as mpl
mpl.use('TkAgg')
%matplotlib inline
import matplotlib.pyplot as plt

to_use = 'Untransformed'

bos_tracts = pd.read_csv('/Users/ritalaplante/Desktop/Thesis Data and Analytics/Census_2010_Tracts.csv', index_col = 0)
bos_tracts.rename(columns = {'GEOID10':'geoid'}, inplace = True)


dfRF = pd.read_csv('/Users/ritalaplante/Desktop/Thesis Data and Analytics/07-Neighborhood Predictions/Predicted'+to_use+'RF.csv', index_col=0)
dfETR = pd.read_csv('/Users/ritalaplante/Desktop/Thesis Data and Analytics/07-Neighborhood Predictions/Predicted'+to_use+'ETR.csv', index_col=0)
dfBos = pd.read_csv('/Users/ritalaplante/Desktop/Thesis Data and Analytics/07-Neighborhood Predictions/Predicted BostonUntransformedRF.csv', index_col=0)

dfRF_bos = pd.merge(dfRF, bos_tracts, on='geoid', how='inner')
dfETR_bos = pd.merge(dfETR, bos_tracts, on='geoid', how='inner')
dfRF_bos.set_index('geoid', inplace = True)
dfETR_bos.set_index('geoid', inplace = True)

cols = ['SES Percentile Ascent 2010-2019', 'SES Percentile Ascent 2019-2028']

dfRF = dfRF[dfRF.columns.intersection(cols)]
dfETR = dfETR[dfETR.columns.intersection(cols)]
dfBos = dfETR[dfETR.columns.intersection(cols)]

dfRF_bos = dfRF[dfRF.columns.intersection(cols)]
dfETR_bos = dfETR[dfETR.columns.intersection(cols)]

dfBos.head()

cols = list(dfETR.columns)

for col in cols:
    col_zscore = col + '_zscore'
    
    dfETR[col_zscore] = (dfETR[col] - dfETR[col].mean())/dfETR[col].std(ddof=0)
    dfRF[col_zscore] = (dfRF[col] - dfRF[col].mean())/dfRF[col].std(ddof=0)
    dfBos[col_zscore] = (dfBos[col] - dfBos[col].mean())/dfBos[col].std(ddof=0)
    
    dfETR_bos[col_zscore] = (dfETR_bos[col] - dfETR_bos[col].mean())/dfETR_bos[col].std(ddof=0)
    dfRF_bos[col_zscore] = (dfRF_bos[col] - dfRF_bos[col].mean())/dfRF_bos[col].std(ddof=0)

dfBos.head()

# dfRF.to_csv('/Users/ritalaplante/Desktop/Thesis Data and Analytics/07-Neighborhood Predictions/' + to_use + 'RFStandardized.csv', index=True)
# dfETR.to_csv('/Users/ritalaplante/Desktop/Thesis Data and Analytics/07-Neighborhood Predictions/' + to_use + 'ETRStandardized.csv', index=True)

# dfRF_bos.to_csv('/Users/ritalaplante/Desktop/Thesis Data and Analytics/07-Neighborhood Predictions/' + to_use + 'RFStandardizedBos.csv', index=True)
# dfETR_bos.to_csv('/Users/ritalaplante/Desktop/Thesis Data and Analytics/07-Neighborhood Predictions/' + to_use + 'ETRStandardizedBos.csv', index=True)

dfBos.to_csv('/Users/ritalaplante/Desktop/Thesis Data and Analytics/07-Neighborhood Predictions/' + to_use + 'StandardizedBoston.csv', index=True)

fig = plt.figure('SES_PR_ASC 2010-2019')
sns.distplot(dfETR_bos['SES Percentile Ascent 2010-2019'], kde=True)      
fig = plt.gcf() # *G*et the *C*urrent *F*igure environment so that the next command works
plt.savefig('/Users/ritalaplante/Desktop/Thesis Data and Analytics/08-Plot Outputs/SES Ascent 2010-2019-ScoreDistETRBos.png', bbox_inches="tight")
plt.close()
print("Done.")

fig = plt.figure('SES_PR_ASC 2010-2019')
sns.distplot(dfRF_bos['SES Percentile Ascent 2010-2019'], kde=True)      
fig = plt.gcf() # *G*et the *C*urrent *F*igure environment so that the next command works
plt.savefig('/Users/ritalaplante/Desktop/Thesis Data and Analytics/08-Plot Outputs/SES Ascent 2010-2019-ScoreDistRFBos.png', bbox_inches="tight")
plt.close()
print("Done.")

fig = plt.figure('SES_PR_ASC 2019-2028')
sns.distplot(dfETR_bos['SES Percentile Ascent 2019-2028'], kde=True)      
fig = plt.gcf() # *G*et the *C*urrent *F*igure environment so that the next command works
plt.savefig('/Users/ritalaplante/Desktop/Thesis Data and Analytics/08-Plot Outputs/SES Ascent 2019-2028-ScoreDistETRBos.png', bbox_inches="tight")
plt.close()
print("Done.")

fig = plt.figure('SES_PR_ASC 2019-2028')
sns.distplot(dfRF_bos['SES Percentile Ascent 2019-2028'], kde=True)      
fig = plt.gcf() # *G*et the *C*urrent *F*igure environment so that the next command works
plt.savefig('/Users/ritalaplante/Desktop/Thesis Data and Analytics/08-Plot Outputs/SES Ascent 2019-2028-ScoreDistRFBos.png', bbox_inches="tight")
plt.close()
print("Done.")
